+++ CHANGELOG.md
@@ -15,6 +15,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 * Remove auto entering to the edit mode on cell focus.
 * Row options column menu item click doesn't trigger an action.
 
+## Added
+* Add new option to generation, typeDetection= 'inference' | 'raw'.
+
 ## [9.1.2] - 2020-05-27
 ### BREAKING
 - q-grid is block element by default, was inline.